backend/requirements.txt
fastapi==0.88.0
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
numpy==1.21.6
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
pandas==1.3.5
Pillow==9.3.0
pydantic==1.10.2
#pymongo==3.13.0
pymongo[srv]==3.13.0
pytorch-lightning==1.8.0.post1
roboflow==0.2.18
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.10.2
torchvision==0.14.0
transformers==4.25.1
uvicorn==0.20.0
faker==17.0.0
timm==0.6.12
pyzbar==0.1.9
python-multipart
passlib==1.7.4
PyJWT==2.6.0
bcrypt==4.0.1
alembic==1.10.2
psycopg2-binary==2.9.5